requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-leaflet==0.29.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
gunicorn==22.0.0
packaging==24.1
pillow==10.3.0
psycopg==3.1.9
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
